Refactor home page hero image
