+++ src/components/Header/Header.css
@@ -33,3 +33,8 @@
     width: 60%;
     justify-content: space-between
 }
+
+.nav-link {
+    cursor: pointer;
+
+}
