+++ src/components/HomePage/HomePage.css
@@ -32,6 +32,10 @@
     margin-top: 0;
 }
 
+.description-container h2 {
+    font-size: 3em;
+}
+
 .description-container button {
     font-family: 'montserrat';
     font-size: 1.2em;
@@ -41,6 +45,9 @@
     border: none;
     height: 2em;
     width: 13em;
+    margin-bottom: 2em;
+    cursor: pointer;
+
 }
 
 .login-container {
@@ -48,9 +55,16 @@
     flex-direction: column;
     justify-content: center;
     align-items: center;
-    background-image: url('../../frame.png');
+    background-image: url('../../assets/paint.png');
     background-repeat: no-repeat;
     background-position: center;
-    height: 65%;
+    background-size: 29em;
+    height: 70%;
     width: 100%;
+}
+
+.login-container h2 {
+    color: white;
+    background: black;
+    font-size: 1.8em;
 }
